+++ src/steps/index.ts
@@ -1,6 +1,15 @@
-import { accountSteps } from './account';
-import { accessSteps } from './access';
+import { tenantSteps } from './tenant';
+import { deviceSteps } from './device';
+import { userSteps } from './user';
+import { userConfigurationSteps } from './user-configuration';
+import { appInstanceSteps } from './app-instance';
 
-const integrationSteps = [...accountSteps, ...accessSteps];
+const integrationSteps = [
+  ...tenantSteps,
+  ...deviceSteps,
+  ...userSteps,
+  ...userConfigurationSteps,
+  ...appInstanceSteps,
+];
 
 export { integrationSteps };